feat(serial): implement a simple serial protocol
This is a very simple protocol to control the motors and receive
information back. The protocol assumes that it will receive a command of the form
flag left right, or simply a flag in the case of sending back information such
as odometry. A few flags were defined, which can be found in the
serial_protocol.hpp.